backend/ml-services/training/requirements.txt
# AWS SDK
boto3>=1.34.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
transformers>=4.35.0
torch>=2.1.0
torch-geometric>=2.4.0
torchvision>=0.16.0
scikit-learn>=1.3.0
datasets>=2.14.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
